Extract common app functionality into a mixin and add elements to the createInstance page to resemble a standalone page
